Devex 827: refactor environment specific login functions to a factory
