+++ cypress-smoketests/integration/case-migration.spec.js
@@ -5,26 +5,16 @@ const {
   CASE_WITH_OTHER_PETITIONERS,
 } = require('../fixtures/caseMigrations');
 const {
-  getRestApi: getRestApiDeployed,
-  getUserToken: getUserTokenDeployed,
-} = require('../support/pages/login');
-const {
-  getRestApi: getRestApiLocal,
-  getUserToken: getUserTokenLocal,
-} = require('../support/pages/local-login');
-
-const SMOKETESTS_LOCAL = Cypress.env('SMOKETESTS_LOCAL');
-
-const getRestApi = SMOKETESTS_LOCAL ? getRestApiLocal : getRestApiDeployed;
-const getUserToken = SMOKETESTS_LOCAL
-  ? getUserTokenLocal
-  : getUserTokenDeployed;
+  getEnvironmentSpecificFunctions,
+} = require('../support/pages/environment-specific-factory');
 
 let token = null;
 let restApi = null;
 
 describe('Case Migration', () => {
   let docketNumber;
+
+  const { getRestApi, getUserToken } = getEnvironmentSpecificFunctions();
 
   before(async () => {
     const results = await getUserToken(
