+++ cypress-smoketests/integration/court-issued-documents.spec.js
@@ -12,15 +12,6 @@ const {
   uploadCourtIssuedDocPdf,
 } = require('../support/pages/case-detail');
 const {
-  closeScannerSetupDialog: closeScannerSetupDialogDeployed,
-  goToCreateCase,
-  goToReviewCase,
-  serveCaseToIrs,
-} = require('../support/pages/create-paper-case');
-const {
-  closeScannerSetupDialog: closeScannerSetupDialogLocal,
-} = require('../support/pages/create-paper-case-local');
-const {
   completeWizardStep1,
   completeWizardStep2,
   completeWizardStep3,
@@ -40,30 +31,25 @@ const {
   fillInCreateCaseFromPaperForm,
 } = require('../../cypress/support/pages/create-paper-petition');
 const {
-  getUserToken: getUserTokenDeployed,
-  login: loginDeployed,
-} = require('../support/pages/login');
+  getEnvironmentSpecificFunctions,
+} = require('../support/pages/environment-specific-factory');
 const {
-  getUserToken: getUserTokenLocal,
-  login: loginLocal,
-} = require('../support/pages/local-login');
-
-const SMOKETESTS_LOCAL = Cypress.env('SMOKETESTS_LOCAL');
-
-const getUserToken = SMOKETESTS_LOCAL
-  ? getUserTokenLocal
-  : getUserTokenDeployed;
-const login = SMOKETESTS_LOCAL ? loginLocal : loginDeployed;
-const closeScannerSetupDialog = SMOKETESTS_LOCAL
-  ? closeScannerSetupDialogLocal
-  : closeScannerSetupDialogDeployed;
-
+  goToCreateCase,
+  goToReviewCase,
+  serveCaseToIrs,
+} = require('../support/pages/create-paper-case');
 const { goToMyDocumentQC } = require('../support/pages/document-qc');
 
 let token = null;
 const testData = {};
 
 const DEFAULT_ACCOUNT_PASS = Cypress.env('DEFAULT_ACCOUNT_PASS');
+
+const {
+  closeScannerSetupDialog,
+  getUserToken,
+  login,
+} = getEnvironmentSpecificFunctions();
 
 describe('Petitioner', () => {
   before(async () => {
